feat(web): settings goes route-backed tabs — Organization | Members | Data & privacy | Notifications (user-ratified 2026-07-20)
/dashboard/settings is now a layout (title + underline RouteTabs bar,
tablist-of-links with aria-selected/aria-current + roving tabindex) over
four REAL sub-routes; the bare route redirect()s to /organization (live
IA). Sections moved intact from the single-card page: org profile +
Appearance tail → /organization, members & roles → /members, AI consent
+ bank links + export card above the bordered danger card (MI-15 purge
modal unchanged) → /data-privacy, reminder toggles → /notifications.
AppNav's Settings entry matches nested sub-routes; the tab bar h-scrolls
inside the viewport below md. Settings-navigating e2e migrated to the
sub-routes + a routed-tabs journey spec; RouteTabs unit-tested; §4 route
map + as-built notes updated.
