+++ web/e2e/mobile-responsive.spec.ts
@@ -120,7 +120,12 @@ const PERSONAL_ROUTES = [
   "/dashboard/accounts",
   "/dashboard/reports",
   "/dashboard/categories",
-  "/dashboard/settings",
+  // routed settings tabs (2026-07-20): the bar itself must h-scroll in
+  // the viewport, and every pane reflows — sweep all four sub-routes.
+  "/dashboard/settings/organization",
+  "/dashboard/settings/members",
+  "/dashboard/settings/data-privacy",
+  "/dashboard/settings/notifications",
 ];
 
 const COMPANY_ROUTES = [
