+++ web/src/app/dashboard/settings/data-privacy/page.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { PrivacySection } from "../sections";
+
+/**
+ * B9 — Data & privacy tab (user-ratified 2026-07-20, B9b frame): AI
+ * consent + bank-link permissions, the USR-001 export card (determinate
+ * strip) above the bordered danger card (USR-002 purge, MI-15 typed
+ * confirm + 7-day grace + cancel).
+ */
+export default function DataPrivacySettingsPage() {
+  return <PrivacySection />;
+}
