+++ web/src/app/dashboard/settings/members/page.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { MembersSection } from "../sections";
+
+/**
+ * B9 — Members tab (user-ratified 2026-07-20): members & roles; email
+ * invites stay pending until that email's first sign-in.
+ */
+export default function MembersSettingsPage() {
+  return <MembersSection />;
+}
