+++ web/src/app/dashboard/settings/notifications/page.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { NotificationsSection } from "../sections";
+
+/**
+ * B9 — Notifications tab (user-ratified 2026-07-20): deadline reminders
+ * + import summaries.
+ */
+export default function NotificationsSettingsPage() {
+  return <NotificationsSection />;
+}
