+++ web/src/app/dashboard/settings/page.tsx
@@ -1,7 +1,11 @@
-import React from "react";
-import SettingsView from "./SettingsView";
+import { redirect } from "next/navigation";
+import { FIRST_SETTINGS_TAB } from "./tabs";
 
-/** B9 `/dashboard/settings` — Settings incl. rights (pages.md B9). */
+/**
+ * B9 `/dashboard/settings` — live IA (user-ratified 2026-07-20): the
+ * bare route lands on the first settings tab; every section is a routed
+ * tab at /dashboard/settings/<tab>.
+ */
 export default function SettingsPage() {
-  return <SettingsView />;
+  redirect(FIRST_SETTINGS_TAB);
 }
